src/save_data/requirements.txt
kafka-python
sqlalchemy
psycopg2
pydantic